[Feature #160840029] Add a router to delete an order of a given id
