--- .env.example
@@ -1,1 +1,0 @@
-PORT=8080
