+++ .travis.yml
@@ -1,6 +1,8 @@
 language: node_js
 node_js: 
   - "stable"
+before_script:
+  - cd server && npm install
 cache:
   directories:
     - "node_modules"
