--- Procfile
@@ -1,1 +1,0 @@
-web: node build/src/index.js
