--- package.json
@@ -1,88 +1,0 @@
-{
-  "name": "fast-food-fast",
-  "version": "1.0.0",
-  "description": "An API for a food Ordering website (Fast-Food-Fast).",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/vkingmaker/fast-food-fast-api.git"
-  },
-  "author": "Monday Victor Akubudike",
-  "license": "MIT",
-  "main": "build/src/index.js",
-  "engines": {
-    "node": "10.11.0",
-    "npm": "6.4.1"
-  },
-  "scripts": {
-    "start": "node build/src/app.js",
-    "clean": "rimraf dist",
-    "build": "babel src/**/*.js --out-dir build --presets=es2015",
-    "test": "mocha test/*.js",
-    "lint": "esw -w src test"
-  },
-  "keywords": [
-    "express",
-    "babel",
-    "es6",
-    "es2015",
-    "es2016",
-    "es2017",
-    "jest",
-    "eslint",
-    "Mocha",
-    "Chai",
-    "Andela"
-  ],
-  "dependencies": {
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.24.1",
-    "body-parser": "^1.18.2",
-    "ejs": "^2.6.1",
-    "express": "^4.16.2",
-    "morgan": "^1.9.0",
-    "rimraf": "^2.6.2"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.0.3",
-    "babel-plugin-add-module-exports": "^1.0.0",
-    "babel-register": "^6.26.0",
-    "chai": "^3.5.0",
-    "chai-http": "^2.0.1",
-    "dotenv": "^4.0.0",
-    "eslint": "^4.12.1",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jest": "^21.3.2",
-    "eslint-watch": "^3.1.3",
-    "mocha": "^2.4.5",
-    "nodemon": "^1.12.1"
-  },
-  "babel": {
-    "presets": [
-      "es2015"
-    ],
-    "plugins": [
-      "transform-object-rest-spread",
-      "transform-class-properties",
-      "add-module-exports"
-    ]
-  },
-  "eslintConfig": {
-    "parser": "babel-eslint",
-    "plugins": [
-      "import"
-    ],
-    "parserOptions": {
-      "ecmaVersion": 2017,
-      "sourceType": "module"
-    },
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "eslint:recommended"
-    ]
-  }
-}
