--- public/styles/style.css
@@ -1,8 +1,0 @@
-body {
-  padding: 50px;
-  font: 14px sans-serif;
-}
-
-a {
-  color: #00B7FF;
-}
