+++ server/.env.example
@@ -1,0 +1,1 @@
+PORT=8080
