+++ server/.travis.yml
@@ -1,0 +1,6 @@
+language: node_js
+node_js: 
+  - "stable"
+cache:
+  directories:
+    - "node_modules"
