+++ server/Procfile
@@ -1,0 +1,1 @@
+web: node build/src/index.js
