+++ server/README.md
@@ -1,0 +1,5 @@
+# fast-food-fast-api
+This API is developed for a food ordering web application. On this plat form you can be able to order for food using The /POST end point, You can also get the list of the whole order saved (its not parsistent since its a JSON file that saves the orders in datastructure) usin the /GET endpoint.
+The user of this API can also get a specific order using the ID of that particular order on /GET/ID endpoint.
+/PUT/ID for updating a particular order,
+/DELETE to clear the data structure id need be of /DELETE/ID to delete a specific order
