+++ server/public/styles/style.css
@@ -1,0 +1,8 @@
+body {
+  padding: 50px;
+  font: 14px sans-serif;
+}
+
+a {
+  color: #00B7FF;
+}
