--- src/app.js
@@ -1,43 +1,0 @@
-import express from 'express';
-import path from 'path';
-import morgan from 'morgan';
-import bodyParser from 'body-parser';
-import routes from './routes';
-
-const app = express();
-
-// View engine setup
-app.set('views', path.join(__dirname, '../views'));
-app.set('view engine', 'ejs');
-
-app.use(morgan('dev'));
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(express.static(path.join(__dirname, '../public')));
-
-// Routes
-app.use('/v1/order', routes);
-
-// Catch 404 and forward to error handler
-app.use((req, res, next) => {
-  const err = new Error('Not Found');
-  err.status = 404;
-  next(err);
-});
-
-// Error handler
-app.use((err, req, res, next) => { // eslint-disable-line no-unused-vars
-  res
-    .status(err.status || 500)
-    .render('error', {
-      message: err.message
-    });
-});
-
-
-// export default app;
-
-const { PORT = 3000 } = process.env;
-app.listen(PORT, () => console.log(`Listening on port ${PORT}`));
-
-export default app;
