--- src/mock/db.js
@@ -1,48 +1,0 @@
- module.exports = function () {
-     return [{
-         "id": 1,
-         "Date": "8/30/2018",
-         "Name": "Monday Victor Akubudike",
-         "Address": "7b Oyadiran Estate Sabo, Yaba",
-         "Tel.No": "07031977216",
-         "Order": "Fried Rice with chicken",
-         "Instruction": "I love my chicken spicy and grilled",
-         "status": "pending"
-     }, {
-         "id": 2,
-         "Date": "8/30/2018",
-         "Name": "Ufomba Victoria",
-         "Address": "1b Nweke street Aba, Abia State",
-         "Tel.No": "07061277913",
-         "Order": "Nsala and fufu",
-         "Instruction": "should be served HOT",
-         "status": "Accepted"
-     }, {
-         "id": 3,
-         "Date": "8/30/2018",
-         "Name": "Monday Otti M",
-         "Address": "4 Ehiemere Crescent, Faulks Road Aba Abia state",
-         "Tel.No": "0813458277",
-         "Order": "Fufu and Okra Soup",
-         "Instruction": "I will like it to be served with fried fish",
-         "status": "Completed"
-     }, {
-         "id": 4,
-         "Date": "8/30/2018",
-         "Name": "Otti Grant Onyeukwu",
-         "Address": "3 Olonode street, Alagumeji Yaba, Lagos state",
-         "Tel.No": "081114537",
-         "Order": "Chiken and Chips",
-         "Instruction": "the ketchup has to be enough",
-         "status": "accepted"
-     }, {
-         "id": 5,
-         "Date": "8/30/2018",
-         "Name": "Tarmalily Cooker",
-         "Address": "11 omenazu street Umungasi Aba abia state",
-         "Tel.No": "08031295499",
-         "Order": "Fried Rice with chicken",
-         "Instruction": "Please deliver my food in Time this time around",
-         "status": "declined"
-     }]
- }
